requirements.txt
transformers==4.51.3
datasets
accelerate
deepspeed
hydra-core
jsonlines
nltk
numpy
omegaconf
peft
PyYAML
rouge_score
scikit_learn
scipy
sentence_transformers
tqdm
bitsandbytes
